feat: criado  endpoints de crud para cidade, endereco e lotacao.

+++ src/main/java/com/seplag/employee_manager/application/io/EnderecoRequest.java
@@ -3,7 +3,7 @@ package com.seplag.employee_manager.application.io;
 public record EnderecoRequest(
     String tipoLogradouro,
     String logradouro,
-    Integer numero,
+    Long numero,
     String bairro,
     CidadeRequest cidade
 ){}
